requirements.txt
fastapi~=0.61.1
uvicorn~=0.12.1
tensorflow~=2.3.1
numpy~=1.18.1
Pillow~=7.2.0
pydantic~=1.6.1